requirements.txt
imageio==2.34.1
matplotlib~=3.9.2
opencv-python~=4.8.0.76
PyYAML~=6.0.2
scikit-learn~= 1.5.2
scipy~=1.11.4
iopath==0.1.10
tqdm
numpy~=1.24.4
typing_extensions~=4.11.0
Pillow~=9.4.0
hydra-core~=1.3.2
tensorboardX~=2.6.2.2
setuptools~=69.5.1
timm~=0.4.12
attrs~=23.2.0
einops==0.8.1
ttach